requirements.txt
gradio>=4.0.0
opencv-python>=4.5.5
numpy>=1.21.0
torch>=2.1.2
Pillow>=10.0.0
flash-attn==2.7.3 